requirements.txt
flask==2.3.3
flask-wtf==1.1.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
numpy==1.25.2
matplotlib==3.7.2
plotly==5.16.1
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.33.1
torch==2.0.1
scikit-learn==1.3.0
gunicorn==21.2.0
pytest==7.4.0
atproto==0.0.33 